making tab animated
