+++ lib/tab_switcher.dart
@@ -1,3 +1,4 @@
 export 'utils/controller.dart';
 export 'utils/impl.dart';
 export 'utils/theme.dart';
+export 'utils/tab.dart';
