+++ lib/utils/impl.dart
@@ -5,6 +5,7 @@ import '../widgets/preview_capturer_widget.dart';
 import '../widgets/app_bar.dart';
 import 'controller.dart';
 import '../widgets/tab_grid.dart';
+import 'tab.dart';
 import 'theme.dart';
 
 typedef TabWidgetBuilder = Widget Function(
