+++ lib/widgets/app_bar.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 
 import '../utils/controller.dart';
 import '../utils/responsive_page_view_scroll_physics.dart';
+import '../utils/tab.dart';
 import '../utils/theme.dart';
 import 'tab_count_icon.dart';
